prepare for codegen
